Question Types fix

+++ CollegeQuizWeb/Entities/QuestionEntity.cs
@@ -29,7 +29,7 @@ public class QuestionEntity : AbstractAuditableEntity
 
     [ForeignKey(nameof(QuestionTypeEntity))]
     [Column("question_type_id")]
-    public int QuestionType { get; set; }
+    public long QuestionType { get; set; }
     public virtual QuizEntity QuizEntity { get; set; }
     public virtual QuestionTypeEntity QuestionTypeEntity { get; set; }
 }
